# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Service API",
        "description": "API documentation for Translation Service API application.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/translations/{record}": {
            "get": {
                "tags": [
                    "update translation"
                ],
                "summary": "update translation record",
                "operationId": "58eb82cf85965588a949cc7987ef12b1",
                "responses": {
                    "204": {
                        "description": "update translation record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "delete translation"
                ],
                "summary": "Delete translation record",
                "operationId": "4676501f7634120b6799c6976f9b1f57",
                "parameters": [
                    {
                        "name": "record",
                        "in": "path",
                        "description": "ID of the translation record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation record deleted"
                    },
                    "404": {
                        "description": "Translation record not found"
                    }
                }
            }
        }
    },
...
}